myapi/requirements.txt
anyio==4.8.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.1
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
dj-database-url==1.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.53.1
fqdn==1.5.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy>=1.25.0
openpyxl==3.1.5
oscrypto==1.3.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pyparsing==3.1.2
pypdf==5.3.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==3.2.1
pytz==2025.1
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==8.0
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
Send2Trash==1.8.3
setuptools==75.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlmap==1.9.3
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3
uri-template==1.3.0
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.29.1
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xhtml2pdf==0.2.16
